Rename images relationship in InterestItem - Add TrainRideStop schema
